Better naming and location
