+++ flytekit/core/python_function_task.py
@@ -6,7 +6,8 @@ from typing import Any, Callable, Dict, List, Optional, TypeVar, Union
 
 from flytekit.common.tasks.raw_container import _get_container_definition
 from flytekit.core.base_task import PythonTask
-from flytekit.core.context_manager import ExecutionState, FlyteContext, ImageConfig, InstanceVar, SerializationSettings
+from flytekit.core.context_manager import ExecutionState, FlyteContext, ImageConfig, SerializationSettings, \
+    TaskResolverMixin
 from flytekit.core.interface import transform_signature_to_interface
 from flytekit.core.resources import Resources, ResourceSpec
 from flytekit.core.workflow import Workflow, WorkflowFailurePolicy, WorkflowMetadata, WorkflowMetadataDefaults
@@ -159,38 +160,6 @@ def istestfunction(func) -> bool:
     return False
 
 
-class TaskLoader(object):
-    """
-    A TaskLoader that can be used to load the task itself from the actual argument that is captured.
-    The argument itself should be discoverable through the class loading framework.
-    """
-
-    @abstractmethod
-    def name(self) -> str:
-        pass
-
-    @abstractmethod
-    def load_task(self, loader_args: List[str]) -> "PythonInstanceTask":
-        """
-        Given the set of identifier keys, should return one Python Task or raise an error if not found
-        """
-        pass
-
-    @abstractmethod
-    def loader_args(self, var: InstanceVar, for_task: "PythonInstanceTask") -> List[str]:
-        """
-        Return a list of strings that can help identify the parameter Task
-        """
-        pass
-
-    @abstractmethod
-    def get_all_tasks(self) -> List["PythonInstanceTask"]:
-        """
-         Future proof method. Just making it easy to access all tasks (Not required today as we auto register them)
-        """
-        pass
-
-
 class PythonInstanceTask(PythonAutoContainerTask[T], ABC):
     """
     This class should be used as the base class for all Tasks that do not have a user defined function body, but have
@@ -211,25 +180,25 @@ class PythonInstanceTask(PythonAutoContainerTask[T], ABC):
         name: str,
         task_config: T,
         task_type: str = "python-task",
-        task_loader: Optional[TaskLoader] = None,
+        task_resolver: Optional[TaskResolverMixin] = None,
         **kwargs,
     ):
         super().__init__(name=name, task_config=task_config, task_type=task_type, **kwargs)
-        self._task_loader = task_loader
+        self._task_resolver = task_resolver
 
     @property
-    def task_loader(self):
-        return self._task_loader
+    def task_resolver(self) -> Optional[TaskResolverMixin]:
+        return self._task_resolver
 
     def get_command(self, settings: SerializationSettings) -> List[str]:
         """
         NOTE: This command is different, it tries to retrieve the actual LHS of where this object was assigned, so that
         the module loader can easily retreive this for execution - at runtime.
         """
-        if self.task_loader is None:
+        if self.task_resolver is None:
             var = settings.get_instance_var(self)
         else:
-            var = settings.get_instance_var(self.task_loader)
+            var = settings.get_instance_var(self.task_resolver)
 
         if var is None:
             raise AssertionError(f"Unable to load instance of task {self.name}, should be loadable in module")
@@ -248,8 +217,8 @@ class PythonInstanceTask(PythonAutoContainerTask[T], ABC):
             "{{.rawOutputDataPrefix}}",
         ]
 
-        if self.task_loader is not None:
-            for arg in self.task_loader.loader_args(var=var, for_task=self):
+        if self.task_resolver is not None:
+            for arg in self.task_resolver.loader_args(var=var, for_task=self):
                 command.extend(["--loader-arg", arg])
 
         return command
@@ -284,7 +253,6 @@ class PythonFunctionTask(PythonInstanceTask[T]):
         task_type="python-task",
         ignore_input_vars: Optional[List[str]] = None,
         execution_mode: Optional[ExecutionBehavior] = ExecutionBehavior.DEFAULT,
-        allow_nested: bool = False,
         **kwargs,
     ):
         """
@@ -296,13 +264,6 @@ class PythonFunctionTask(PythonInstanceTask[T]):
         """
         if task_function is None:
             raise ValueError("TaskFunction is a required parameter for PythonFunctionTask")
-        if not allow_nested:
-            if not istestfunction(func=task_function) and isnested(func=task_function):
-                raise ValueError(
-                    "TaskFunction cannot be a nested/inner or local function. "
-                    "It should be accessible at a module level for Flyte to execute it. Test modules with "
-                    "names begining with `test_` are allowed to have nested tasks"
-                )
         self._native_interface = transform_signature_to_interface(inspect.signature(task_function))
         mutated_interface = self._native_interface.remove_inputs(ignore_input_vars)
         super().__init__(
@@ -312,6 +273,14 @@ class PythonFunctionTask(PythonInstanceTask[T]):
             task_config=task_config,
             **kwargs,
         )
+        if self.task_resolver is None:
+            if not istestfunction(func=task_function) and isnested(func=task_function):
+                raise ValueError(
+                    "TaskFunction cannot be a nested/inner or local function. "
+                    "It should be accessible at a module level for Flyte to execute it. Test modules with "
+                    "names begining with `test_` are allowed to have nested tasks. If you want to create your own tasks"
+                    "use the TaskResolverMixin"
+                )
         self._task_function = task_function
         self._execution_mode = execution_mode
 
@@ -334,7 +303,7 @@ class PythonFunctionTask(PythonInstanceTask[T]):
             return self.dynamic_execute(self._task_function, **kwargs)
 
     def get_command(self, settings: SerializationSettings) -> List[str]:
-        if self.task_loader is not None:
+        if self.task_resolver is not None:
             return super().get_command(settings)
         return [
             "pyflyte-execute",
